Trait bit_utils::BitInformation

pub trait BitInformation {
    fn number_of_bits(&self) -> usize;
    fn has_x_bit(&self, position: usize) -> bool;

    fn is_bit_in_bounds(&self, position: usize) -> bool { ... }
    fn has_most_signifigant_bit(&self) -> bool { ... }
}

Required Methods

fn number_of_bits(&self) -> usize

Gets the number of bits contained in this type

fn has_x_bit(&self, position: usize) -> bool

Checks to see if bit X is set

Provided Methods

fn is_bit_in_bounds(&self, position: usize) -> bool

Checks to see if the requested bit position is in bounds

fn has_most_signifigant_bit(&self) -> bool

Checks to see if the most signifigant bit is set